Remove cheap hack on node name
Seems the jenkins remote executor, despite being configured to connect
to the new external verifudge box, was still connected to the internal
verifudge box. By restarting it the deployments are now happening on the
correct box.

I've removed this node name hack to see whether this is now
no longer needed. I'm guessing the symtoms I saw were due to the
executor being connected to the wrong box, not the node name not being
listened to.